+++ noodles-bam/src/writer.rs
@@ -298,7 +298,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use noodles_sam::{record::Data, AlignmentRecord, AlignmentWriter};
+    use noodles_sam::{record::Data, AlignmentWriter};
 
     use super::*;
     use crate::Reader;