+++ noodles-sam/examples/sam_filter_data.rs
@@ -8,7 +8,7 @@ use std::{
     io::{self, BufReader},
 };
 
-use noodles_sam::{self as sam, alignment::Record, record::data::field::Tag, AlignmentRecord};
+use noodles_sam::{self as sam, alignment::Record, record::data::field::Tag};
 
 fn is_unique_record(record: &Record) -> io::Result<bool> {
     match record.data().get(Tag::AlignmentHitCount) {