+++ noodles-sam/src/alignment/record.rs
@@ -9,7 +9,7 @@ use noodles_core::Position;
 use crate::{
     header::{ReferenceSequence, ReferenceSequences},
     record::{Cigar, Data, Flags, MappingQuality, QualityScores, ReadName, Sequence},
-    AlignmentRecord,
+    Header,
 };
 
 /// An alignment record.
@@ -35,9 +35,19 @@ impl Record {
         Builder::default()
     }
 
+    /// Returns the read name.
+    pub fn read_name(&self) -> Option<&ReadName> {
+        self.read_name.as_ref()
+    }
+
     /// Returns a mutable reference to the read name.
     pub fn read_name_mut(&mut self) -> &mut Option<ReadName> {
         &mut self.read_name
+    }
+
+    /// Returns the flags.
+    pub fn flags(&self) -> Flags {
+        self.flags
     }
 
     /// Returns a mutable reference to the flags.
@@ -55,14 +65,29 @@ impl Record {
         &mut self.reference_sequence_id
     }
 
+    /// Returns the alignment start.
+    pub fn alignment_start(&self) -> Option<Position> {
+        self.alignment_start
+    }
+
     /// Returns a mutable reference to the alignment start.
     pub fn alignment_start_mut(&mut self) -> &mut Option<Position> {
         &mut self.alignment_start
     }
 
+    /// Returns the mapping quality.
+    pub fn mapping_quality(&self) -> Option<MappingQuality> {
+        self.mapping_quality
+    }
+
     /// Returns a mutable reference to the mapping quality.
     pub fn mapping_quality_mut(&mut self) -> &mut Option<MappingQuality> {
         &mut self.mapping_quality
+    }
+
+    /// Returns the CIGAR operations.
+    pub fn cigar(&self) -> &Cigar {
+        &self.cigar
     }
 
     /// Returns a mutable reference to the CIGAR operations.
@@ -80,9 +105,19 @@ impl Record {
         &mut self.mate_reference_sequence_id
     }
 
+    /// Returns the mate alignment start.
+    pub fn mate_alignment_start(&self) -> Option<Position> {
+        self.mate_alignment_start
+    }
+
     /// Returns a mutable reference to the mate alignment start.
     pub fn mate_alignment_start_mut(&mut self) -> &mut Option<Position> {
         &mut self.mate_alignment_start
+    }
+
+    /// Returns the template length.
+    pub fn template_length(&self) -> i32 {
+        self.template_length
     }
 
     /// Returns a mutable reference to template length.
@@ -90,9 +125,19 @@ impl Record {
         &mut self.template_length
     }
 
+    /// Returns the sequence.
+    pub fn sequence(&self) -> &Sequence {
+        &self.sequence
+    }
+
     /// Returns a mutable reference to sequence.
     pub fn sequence_mut(&mut self) -> &mut Sequence {
         &mut self.sequence
+    }
+
+    /// Returns the quality scores.
+    pub fn quality_scores(&self) -> &QualityScores {
+        &self.quality_scores
     }
 
     /// Returns a mutable reference to quality scores.
@@ -100,75 +145,67 @@ impl Record {
         &mut self.quality_scores
     }
 
+    /// Returns the data.
+    pub fn data(&self) -> &Data {
+        &self.data
+    }
+
     /// Returns a mutable reference to the data.
     pub fn data_mut(&mut self) -> &mut Data {
         &mut self.data
+    }
+
+    /// Returns the associated reference sequence.
+    pub fn reference_sequence<'a>(
+        &self,
+        header: &'a Header,
+    ) -> Option<io::Result<&'a ReferenceSequence>> {
+        get_reference_sequence(header.reference_sequences(), self.reference_sequence_id())
+    }
+
+    /// Returns the associated mate reference sequence.
+    pub fn mate_reference_sequence<'a>(
+        &self,
+        header: &'a Header,
+    ) -> Option<io::Result<&'a ReferenceSequence>> {
+        get_reference_sequence(
+            header.reference_sequences(),
+            self.mate_reference_sequence_id(),
+        )
+    }
+
+    /// Returns the alignment span.
+    pub fn alignment_span(&self) -> usize {
+        self.cigar().alignment_span()
+    }
+
+    /// Calculates the end position.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_core::Position;
+    /// use noodles_sam::{self as sam, alignment::Record};
+    ///
+    /// let record = Record::builder()
+    ///     .set_alignment_start(Position::try_from(8)?)
+    ///     .set_cigar("5M".parse()?)
+    ///     .build();
+    ///
+    /// assert_eq!(record.alignment_end(), Position::new(12));
+    /// # Ok::<_, Box<dyn std::error::Error>>(())
+    /// ```
+    pub fn alignment_end(&self) -> Option<Position> {
+        self.alignment_start().and_then(|alignment_start| {
+            let end = usize::from(alignment_start) + self.alignment_span() - 1;
+            Position::new(end)
+        })
     }
 }
 
 impl Default for Record {
     fn default() -> Self {
         Self::builder().build()
-    }
-}
-
-impl AlignmentRecord for Record {
-    fn read_name(&self) -> Option<&ReadName> {
-        self.read_name.as_ref()
-    }
-
-    fn reference_sequence<'rs>(
-        &self,
-        reference_sequences: &'rs ReferenceSequences,
-    ) -> Option<std::io::Result<&'rs ReferenceSequence>> {
-        get_reference_sequence(reference_sequences, self.reference_sequence_id())
-    }
-
-    fn flags(&self) -> Flags {
-        self.flags
-    }
-
-    fn alignment_start(&self) -> Option<Position> {
-        self.alignment_start
-    }
-
-    fn alignment_span(&self) -> usize {
-        self.cigar().alignment_span()
-    }
-
-    fn mapping_quality(&self) -> Option<MappingQuality> {
-        self.mapping_quality
-    }
-
-    fn cigar(&self) -> &Cigar {
-        &self.cigar
-    }
-
-    fn mate_reference_sequence<'rs>(
-        &self,
-        reference_sequences: &'rs ReferenceSequences,
-    ) -> Option<std::io::Result<&'rs ReferenceSequence>> {
-        get_reference_sequence(reference_sequences, self.mate_reference_sequence_id())
-    }
-
-    fn mate_alignment_start(&self) -> Option<Position> {
-        self.mate_alignment_start
-    }
-
-    fn template_length(&self) -> i32 {
-        self.template_length
-    }
-
-    fn sequence(&self) -> &Sequence {
-        &self.sequence
-    }
-
-    fn quality_scores(&self) -> &QualityScores {
-        &self.quality_scores
-    }
-
-    fn data(&self) -> &Data {
-        &self.data
     }
 }